no more long long

+++ trunk/src/kernel/rational/givratcstor.C
@@ -66,11 +66,11 @@ Rational::Rational(double x) {
         const long shift = 1075-t.u.exponent;
         t.u.exponent = 1076;
         if (shift > 0) {
-            Integer tt( t.u.mantissa+4503599627370496ULL );
+            Integer tt( t.u.mantissa+static_cast<uint64_t>(4503599627370496ULL) );
             num = (x<0.?-tt:tt);
             den = Integer(1)<<shift;
         } else {
-            Integer tt( t.u.mantissa+4503599627370496ULL);
+            Integer tt( t.u.mantissa+static_cast<uint64_t>(4503599627370496ULL));
             tt <<=(-shift);
             num = (x<0.?-tt:tt);
             den = 1;
